pyproject.toml
[project]
name = "vcon-fadapter"
version = "0.1.0"
description = "A file system monitoring adapter that watches for fax images and creates vCon objects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "vcon>=0.1.0",
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "boto3>=1.28.0",
]

[project.scripts]
vcon-fadapter = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "moto[s3]>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fax_adapter"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

